[editor Port, Module] Move Port/Ports logic & styles into Ports.jsx/Ports.pcss. Add form focus states.
