+++ app/src/editor/components/Module.jsx
@@ -11,25 +11,14 @@ import { DragTypes, RunStates } from '../state'
 
 import { Resizer, isModuleResizable } from './Resizer'
 import RenameInput from './RenameInput'
-import Port from './Port'
+import Ports from './Ports'
 
 import styles from './Module.pcss'
 
 class CanvasModule extends React.Component {
     state = {
         isDraggable: true,
-        minPortSize: 0,
         isResizing: false,
-    }
-
-    portRefs = new Map()
-
-    getPortRef = (portId) => {
-        // memoize ref functions
-        if (!this.portRefs.has(portId)) {
-            this.portRefs.set(portId, (el) => this.props.onPort(portId, el))
-        }
-        return this.portRefs.get(portId)
     }
 
     // for disabling dragging when cursor is over interactive controls e.g. inputs
@@ -37,11 +26,6 @@ class CanvasModule extends React.Component {
         this.setState({
             isDraggable,
         })
-    }
-
-    // for resizing all port widths to match longest port value
-    adjustMinPortSize = (minPortSize) => {
-        this.setState({ minPortSize })
     }
 
     /**
@@ -115,25 +99,9 @@ class CanvasModule extends React.Component {
             connectDragSource,
             isDragging,
         } = this.props
-        const { outputs } = module
         const { isDraggable, layout } = this.state
 
-        const inputs = module.params.concat(module.inputs)
-
-        // map inputs and outputs into visual rows
-        const rows = []
-        const maxRows = Math.max(inputs.length, outputs.length)
-        for (let i = 0; i < maxRows; i += 1) {
-            rows.push([inputs[i], outputs[i]])
-        }
-
         const isSelected = module.hash === this.props.selectedModuleHash
-        const portSize = Math.min(module.params.reduce((size, { value, defaultValue }) => (
-            Math.max(size, String(value || defaultValue).length)
-        ), Math.max(4, this.state.minPortSize)), 40)
-
-        // this is the `display: table` equivalent of `<td colspan="3" />`. For alignment.
-        const PortPlaceholder = () => <React.Fragment><div /><div /><div /></React.Fragment>
 
         const maybeConnectDragging = (el) => (
             isDraggable ? connectDragSource(el) : el
@@ -180,28 +148,7 @@ class CanvasModule extends React.Component {
                         <HamburgerIcon />
                     </button>
                 </div>
-                <div className={styles.ports}>
-                    {rows.map((ports) => (
-                        <div key={ports.map((p) => p && p.id).join(',')} className={styles.portRow} role="row">
-                            {ports.map((port, index) => (
-                                /* eslint-disable react/no-array-index-key */
-                                !port ? <PortPlaceholder key={index} /> /* placeholder for alignment */ : (
-                                    <Port
-                                        key={port.id + index}
-                                        port={port}
-                                        onPort={this.props.onPort}
-                                        size={portSize}
-                                        adjustMinPortSize={this.adjustMinPortSize}
-                                        setIsDraggable={this.setIsDraggable}
-                                        canvas={canvas}
-                                        {...api.port}
-                                    />
-                                )
-                                /* eslint-enable react/no-array-index-key */
-                            ))}
-                        </div>
-                    ))}
-                </div>
+                <Ports {...this.props} />
                 {!!isResizable && !isRunning && (
                     /* eslint-disable-next-line jsx-a11y/mouse-events-have-key-events */
                     <Resizer
