+++ app/src/editor/components/Module.pcss
@@ -1,5 +1,3 @@
-/* stylelint-disable no-descending-specificity */
-
 /**
  * Module
  */
@@ -53,267 +51,4 @@
 
 .moduleHeader .optionsButton:focus {
   outline: none;
-}
-
-.ports {
-  display: table;
-  font-family: 'IBM Plex Mono', sans-serif;
-  border-spacing: 4px;
-  padding: 8px 2px;
-  flex: 1;
-}
-
-.ports .portRow {
-  display: table-row;
-  line-height: 1;
-}
-
-.ports .portRow > * {
-  display: table-cell;
-  padding: 0 2px;
-}
-
-.ports .portName {
-  z-index: 1;
-  position: relative;
-  color: #525252;
-  font-size: 10px;
-  font-weight: 500;
-  white-space: nowrap;
-}
-
-.ports .portName.isOutput {
-  text-align: right;
-  width: 100%;
-}
-
-/**
- * Port Icon
- */
-
-.ports .portRow .portIconContainer {
-  /* rows with inputs change icon vertical alignment
-     add manual offset with padding to counter */
-  padding-top: 0;
-  padding-bottom: 0.25em;
-  width: 8px; /* explicitly set width to ensure table-cell collapses correctly */
-}
-
-.ports .portIcon {
-  display: inline-block;
-  border: 1px solid #ADADAD;
-  content: ' ';
-  width: 8px;
-  height: 8px;
-  background: #FFFFFF;
-  border-radius: 100%;
-  transition: box-shadow 0.2s;
-  transform-origin: center center;
-  position: relative;
-  box-shadow: 0 0 0 2px rgba(255, 255, 255, 0);
-}
-
-.ports .portIcon.requiresConnection {
-  border-radius: 0;
-}
-
-.ports .portIcon.drivingInput {
-  border-color: #FF5C00;
-}
-
-.ports .portIcon.noRepeat {
-  border-color: #0324FF;
-}
-
-.ports .portIcon:not(.isDragging).dragPortInProgress {
-  border-color: rgba(255, 255, 255, 0);
-  background-color: #C90000;
-}
-
-.ports .portIcon:not(.isDragging).dragPortInProgress.canDrop {
-  background-color: #00C900;
-}
-
-.ports .portIcon:not(.dragPortInProgress):hover {
-  box-shadow: 0 0 0 2px #ADF5FF;
-}
-
-.ports .portIcon.connected {
-  border-color: #2AC437;
-}
-
-.ports .portIcon.connected::after {
-  transition: inherit;
-  content: ' ';
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  width: 2px;
-  height: 2px;
-  background: #2AC437;
-  opacity: 1;
-}
-
-.ports .portIcon.connected.dragPortInProgress::after {
-  opacity: 0;
-}
-
-.ports .portIcon.isOver {
-  transform: scale(2);
-}
-
-.ports .portIcon .portOptions {
-  position: absolute;
-  top: 50%;
-  display: grid;
-  grid-template-columns: auto auto auto;
-  grid-column-gap: 2px;
-  opacity: 0;
-  transition: opacity 0.25s;
-}
-
-.ports .portIcon.isInput .portOptions {
-  right: initial;
-  left: 0;
-  padding-right: 13px;
-  transform: translate(-100%, -50%);
-}
-
-.ports .portIcon.isOutput .portOptions {
-  left: initial;
-  right: 0;
-  padding-left: 13px;
-  transform: translate(100%, -50%);
-}
-
-.ports .portIcon:not(.dragPortInProgress):hover .portOptions {
-  opacity: 1;
-}
-
-.ports .portIcon .portOptions > button[type=button] {
-  appearance: none;
-  border: none;
-  padding: 3px 4px;
-  background: #CDCDCD;
-  border-radius: 2px 2px 2px 2px;
-  color: #FFFFFF;
-  font-size: 10px;
-  font-weight: 500;
-  text-align: center;
-  letter-spacing: 0;
-  line-height: 10px;
-  transition: background 0.25s;
-}
-
-.ports .portIcon .portOptions > .drivingInputOption[value='true'] {
-  background-color: #FF5C00;
-}
-
-.ports .portIcon .portOptions > .initialValueOption[value='true'] {
-  background-color: #BF9595;
-}
-
-.ports .portIcon .portOptions > .noRepeatOption[value='true'] {
-  background-color: #0324FF;
-}
-
-/**
- * Port Default Value
- */
-
-.ports .portValue {
-  min-height: 24px;
-  width: 0;
-}
-
-.ports select.portValue,
-.ports input.portValue {
-  width: auto;
-  color: #525252;
-  font-size: 10px;
-  font-weight: 500;
-  font-family: inherit;
-  background: #F5F5F5;
-  border-radius: 2px;
-  border: none;
-  transition: min-width 0.25s;
-  padding: 2px 8px;
-}
-
-.ports select.portValue {
-  width: 100%;
-  height: 24px; /* select ignores min-height */
-}
-
-.ports input.portValue {
-  cursor: text;
-}
-
-.ports input.portValue::placeholder {
-  color: #ADADAD;
-  font-style: italic;
-}
-
-.ports select.portValue:disabled,
-.ports input.portValue:disabled {
-  cursor: default;
-  color: #A2A2A2;
-}
-
-/**
- * Map-type Parameters
- */
-
-.ports .mapParam {
-  display: grid;
-  grid-template-columns: 1fr 1fr auto;
-  grid-column-gap: 8px;
-  grid-row-gap: 4px;
-  width: auto;
-  color: #323232;
-}
-
-.ports input.portValue,
-.ports select.portValue,
-.ports .mapParam input,
-.ports .mapParam button {
-  transition: background 0.3s, color 0.3s;
-  transition-delay: 0;
-}
-
-.ports .mapParam input {
-  min-width: 100%;
-  text-overflow: ellipsis;
-}
-
-.ports .mapParam button {
-  color: #323232;
-  font-size: inherit;
-  font-family: inherit;
-  background: #F5F5F5;
-  border-radius: 2px;
-  border: none;
-  padding: 2px 8px;
-  appearance: none;
-}
-
-.ports .mapParam:hover input.portValue,
-.ports .mapParam:focus-within input.portValue,
-.ports .mapParam:hover button,
-.ports .mapParam:focus-within button {
-  background: #F5F5F5;
-  color: #323232;
-}
-
-.ports input:not(:hover):not(:focus),
-.ports select:not(:hover):not(:focus),
-.ports .mapParam:not(:hover):not(:focus-within) input,
-.ports .mapParam:not(:hover):not(:focus-within) button {
-  background: none;
-  color: #ADADAD;
-}
-
-.ports .mapParam button > svg {
-  display: block;
 }
